testscripts/reintegrate-from-list/requirements.txt
requests==2.21.0
PyYAML==5.1